+++ tasks/omp/vershinina_a_hoare_sort/include/ops_omp.hpp
