+++ tasks/seq/vershinina_a_hoare_sort_disabled/include/ops_seq.hpp
