+++ tasks/seq/vershinina_a_hoare_sort_disabled/src/ops_seq.cpp
@@ -1,4 +1,4 @@
-#include "seq/vershinina_a_hoare_sort_seq/include/ops_seq.hpp"
+#include "seq/vershinina_a_hoare_sort/include/ops_seq.hpp"
 
 #include <algorithm>
