+++ tasks/stl/vershinina_a_hoare_sort/data/test.txt
@@ -1,0 +1,1 @@
+100
