+++ tasks/stl/vershinina_a_hoare_sort/include/ops_stl.hpp
