+++ tasks/stl/vershinina_a_hoare_sort/src/ops_stl.cpp
@@ -1,16 +1,11 @@
-#include "tbb/vershinina_a_hoare_sort_tbb/include/ops_tbb.hpp"
-
-#include <tbb/tbb.h>
+#include "stl/vershinina_a_hoare_sort/include/ops_stl.hpp"
 
 #include <algorithm>
 #include <cmath>
-#include <core/util/include/util.hpp>
-#include <utility>
+#include <thread>
 #include <vector>
 
-#include "oneapi/tbb/blocked_range.h"
-#include "oneapi/tbb/parallel_for.h"
-#include "oneapi/tbb/task_arena.h"
+#include "core/util/include/util.hpp"
 
 namespace {
 int Partition(double *s_vec, int first, int last) {
@@ -42,16 +37,28 @@ void BatcherMergeBlocksStep(double *left_pointer, int &left_size, double *right_
 
 void BatcherMerge(int thread_input_size, std::vector<double *> &pointers, std::vector<int> &sizes, int par_if_greater) {
   const int batcherthreads = ppc::util::GetPPCNumThreads();
+
+  std::vector<std::thread> threads(batcherthreads);
+
   for (int step = 1, pack = int(pointers.size()); pack > 1; step *= 2, pack /= 2) {
-    tbb::task_arena arena(((thread_input_size / step) > par_if_greater) ? batcherthreads : 1);
-    arena.execute([&] {
-      tbb::parallel_for(tbb::blocked_range<int>(0, pack / 2), [&](const auto &r) {
-        for (int off = r.begin(); off < r.end(); off++) {
-          BatcherMergeBlocksStep(pointers[2 * step * off], sizes[2 * step * off], pointers[(2 * step * off) + step],
-                                 sizes[(2 * step * off) + step]);
-        }
-      });
-    });
+    if ((thread_input_size / step) > par_if_greater) {
+      for (int k = 0; k < pack / 2; ++k) {
+        threads[k] = std::thread(
+            [&](int off) {
+              BatcherMergeBlocksStep(pointers[2 * step * off], sizes[2 * step * off], pointers[(2 * step * off) + step],
+                                     sizes[(2 * step * off) + step]);
+            },
+            k);
+      }
+      for (int k = 0; k < pack / 2; ++k) {
+        threads[k].join();
+      }
+    } else {
+      for (int off = 0; off < pack / 2; ++off) {
+        BatcherMergeBlocksStep(pointers[2 * step * off], sizes[2 * step * off], pointers[(2 * step * off) + step],
+                               sizes[(2 * step * off) + step]);
+      }
+    }
     if ((pack / 2) - 1 == 0) {
       BatcherMergeBlocksStep(pointers[0], sizes[sizes.size() - 1], pointers[pointers.size() - 1],
                              sizes[sizes.size() - 1]);
@@ -62,18 +69,19 @@ void BatcherMerge(int thread_input_size, std::vector<double *> &pointers, std::v
   }
 }
 }  // namespace
-bool vershinina_a_hoare_sort_tbb::TestTaskTBB::PreProcessingImpl() {
+
+bool vershinina_a_hoare_sort_stl::TestTaskSTL::PreProcessingImpl() {
   input_.assign(reinterpret_cast<double *>(task_data->inputs[0]),
                 reinterpret_cast<double *>(task_data->inputs[0]) + task_data->inputs_count[0]);
   return true;
 }
 
-bool vershinina_a_hoare_sort_tbb::TestTaskTBB::ValidationImpl() {
+bool vershinina_a_hoare_sort_stl::TestTaskSTL::ValidationImpl() {
   return (task_data->inputs_count[0] == task_data->outputs_count[0]) && task_data->inputs.size() == 1 &&
          task_data->inputs_count.size() == 1 && task_data->outputs.size() == 1;
 }
 
-bool vershinina_a_hoare_sort_tbb::TestTaskTBB::RunImpl() {
+bool vershinina_a_hoare_sort_stl::TestTaskSTL::RunImpl() {
   int n = int(input_.size());
   if (n <= 1) {
     return true;
@@ -92,19 +100,19 @@ bool vershinina_a_hoare_sort_tbb::TestTaskTBB::RunImpl() {
     sizes[i] = thread_input_size;
   }
   sizes[sizes.size() - 1] += thread_input_remainder_size;
-  tbb::task_arena arena(numthreads);
-  arena.execute([&] {
-    tbb::parallel_for(tbb::blocked_range<int>(0, numthreads, 1), [&pointers, &sizes](const auto &r) {
-      for (int i = r.begin(); i < r.end(); i++) {
-        HoareSort(pointers[i], 0, sizes[i] - 1);
-      }
-    });
-  });
+
+  std::vector<std::thread> threads(numthreads);
+  for (int i = 0; i < numthreads; i++) {
+    threads[i] = std::thread(HoareSort, pointers[i], 0, sizes[i] - 1);
+  }
+  for (int i = 0; i < numthreads; i++) {
+    threads[i].join();
+  }
   BatcherMerge(thread_input_size, pointers, sizes, 32);
   return true;
 }
 
-bool vershinina_a_hoare_sort_tbb::TestTaskTBB::PostProcessingImpl() {
+bool vershinina_a_hoare_sort_stl::TestTaskSTL::PostProcessingImpl() {
   std::ranges::copy(res_, reinterpret_cast<double *>(task_data->outputs[0]));
   return true;
 }
