+++ tasks/tbb/vershinina_a_hoare_sort/perf_tests/main.cpp
@@ -10,7 +10,7 @@
 
 #include "core/perf/include/perf.hpp"
 #include "core/task/include/task.hpp"
-#include "tbb/vershinina_a_hoare_sort_tbb/include/ops_tbb.hpp"
+#include "tbb/vershinina_a_hoare_sort/include/ops_tbb.hpp"
 
 namespace {
 std::vector<double> GetRandomVector(int len) {
