+++ tasks/tbb/vershinina_a_hoare_sort/src/ops_tbb.cpp
@@ -1,11 +1,16 @@
-#include "omp/vershinina_a_hoare_sort_omp/include/ops_omp.hpp"
+#include "tbb/vershinina_a_hoare_sort/include/ops_tbb.hpp"
+
+#include <tbb/tbb.h>
 
 #include <algorithm>
 #include <cmath>
+#include <core/util/include/util.hpp>
 #include <utility>
 #include <vector>
 
-#include "core/util/include/util.hpp"
+#include "oneapi/tbb/blocked_range.h"
+#include "oneapi/tbb/parallel_for.h"
+#include "oneapi/tbb/task_arena.h"
 
 namespace {
 int Partition(double *s_vec, int first, int last) {
@@ -36,12 +41,17 @@ void BatcherMergeBlocksStep(double *left_pointer, int &left_size, double *right_
 }
 
 void BatcherMerge(int thread_input_size, std::vector<double *> &pointers, std::vector<int> &sizes, int par_if_greater) {
+  const int batcherthreads = ppc::util::GetPPCNumThreads();
   for (int step = 1, pack = int(pointers.size()); pack > 1; step *= 2, pack /= 2) {
-#pragma omp parallel for if ((thread_input_size / step) > par_if_greater)
-    for (int off = 0; off < pack / 2; ++off) {
-      BatcherMergeBlocksStep(pointers[2 * step * off], sizes[2 * step * off], pointers[(2 * step * off) + step],
-                             sizes[(2 * step * off) + step]);
-    }
+    tbb::task_arena arena(((thread_input_size / step) > par_if_greater) ? batcherthreads : 1);
+    arena.execute([&] {
+      tbb::parallel_for(tbb::blocked_range<int>(0, pack / 2), [&](const auto &r) {
+        for (int off = r.begin(); off < r.end(); off++) {
+          BatcherMergeBlocksStep(pointers[2 * step * off], sizes[2 * step * off], pointers[(2 * step * off) + step],
+                                 sizes[(2 * step * off) + step]);
+        }
+      });
+    });
     if ((pack / 2) - 1 == 0) {
       BatcherMergeBlocksStep(pointers[0], sizes[sizes.size() - 1], pointers[pointers.size() - 1],
                              sizes[sizes.size() - 1]);
@@ -52,18 +62,18 @@ void BatcherMerge(int thread_input_size, std::vector<double *> &pointers, std::v
   }
 }
 }  // namespace
-bool vershinina_a_hoare_sort_omp::TestTaskOpenMP::PreProcessingImpl() {
+bool vershinina_a_hoare_sort_tbb::TestTaskTBB::PreProcessingImpl() {
   input_.assign(reinterpret_cast<double *>(task_data->inputs[0]),
                 reinterpret_cast<double *>(task_data->inputs[0]) + task_data->inputs_count[0]);
   return true;
 }
 
-bool vershinina_a_hoare_sort_omp::TestTaskOpenMP::ValidationImpl() {
+bool vershinina_a_hoare_sort_tbb::TestTaskTBB::ValidationImpl() {
   return (task_data->inputs_count[0] == task_data->outputs_count[0]) && task_data->inputs.size() == 1 &&
          task_data->inputs_count.size() == 1 && task_data->outputs.size() == 1;
 }
 
-bool vershinina_a_hoare_sort_omp::TestTaskOpenMP::RunImpl() {
+bool vershinina_a_hoare_sort_tbb::TestTaskTBB::RunImpl() {
   int n = int(input_.size());
   if (n <= 1) {
     return true;
@@ -82,16 +92,19 @@ bool vershinina_a_hoare_sort_omp::TestTaskOpenMP::RunImpl() {
     sizes[i] = thread_input_size;
   }
   sizes[sizes.size() - 1] += thread_input_remainder_size;
-
-#pragma omp parallel for
-  for (int i = 0; i < numthreads; i++) {
-    HoareSort(pointers[i], 0, sizes[i] - 1);
-  }
+  tbb::task_arena arena(numthreads);
+  arena.execute([&] {
+    tbb::parallel_for(tbb::blocked_range<int>(0, numthreads, 1), [&pointers, &sizes](const auto &r) {
+      for (int i = r.begin(); i < r.end(); i++) {
+        HoareSort(pointers[i], 0, sizes[i] - 1);
+      }
+    });
+  });
   BatcherMerge(thread_input_size, pointers, sizes, 32);
   return true;
 }
 
-bool vershinina_a_hoare_sort_omp::TestTaskOpenMP::PostProcessingImpl() {
+bool vershinina_a_hoare_sort_tbb::TestTaskTBB::PostProcessingImpl() {
   std::ranges::copy(res_, reinterpret_cast<double *>(task_data->outputs[0]));
   return true;
 }
